+++ packages/lambda/CHANGELOG.md
@@ -1,7 +1,7 @@
-# @tayori/lambda
+# @kotodayori/lambda
 
 ## 0.0.1
 
 ### Minor Changes
 
-- Initial release of `@tayori/lambda`: AWS Lambda adapter for Tayori.
+- Initial release of `@kotodayori/lambda`: AWS Lambda adapter for Kotodayori.